+++ webots_ros2_tiago/launch/robot_bringup_launch.py
@@ -17,7 +17,6 @@
 """Launch Webots and the controller."""
 
 import os
-import pathlib
 import launch
 from launch.substitutions import LaunchConfiguration
 from launch.actions import DeclareLaunchArgument
@@ -28,24 +27,21 @@ from ament_index_python.packages import get_package_share_directory
 from launch.launch_description_sources import PythonLaunchDescriptionSource
 from launch.actions import IncludeLaunchDescription
 from webots_ros2_driver.webots_launcher import WebotsLauncher
+from webots_ros2_driver.webots_controller import WebotsController
 from webots_ros2_driver.wait_for_controller_connection import WaitForControllerConnection
-from webots_ros2_driver.utils import controller_url_prefix
 
 
 def get_ros2_nodes(*args):
     package_dir = get_package_share_directory('webots_ros2_tiago')
     use_rviz = LaunchConfiguration('rviz', default=False)
-    robot_description = pathlib.Path(os.path.join(package_dir, 'resource', 'tiago_bringup_webots.urdf')).read_text()
+    robot_description_path = os.path.join(package_dir, 'resource', 'tiago_bringup_webots.urdf')
     ros2_control_params = os.path.join(package_dir, 'resource', 'ros2_control_bringup.yml')
     use_sim_time = LaunchConfiguration('use_sim_time', default=True)
 
-    tiago_driver = Node(
-        package='webots_ros2_driver',
-        executable='driver',
-        output='screen',
-        additional_env={'WEBOTS_CONTROLLER_URL': controller_url_prefix() + 'Tiago'},
+    tiago_driver = WebotsController(
+        robot_name='Tiago',
         parameters=[
-            {'robot_description': robot_description,
+            {'robot_description': robot_description_path,
              'use_sim_time': use_sim_time},
             ros2_control_params
         ]